+++ servicecatalog_puppet/asset_helpers.py
@@ -1,0 +1,15 @@
+import os
+
+
+def resolve_from_site_packages(what):
+    return os.path.sep.join([
+        os.path.dirname(os.path.abspath(__file__)),
+        what
+    ])
+
+
+def read_from_site_packages(what):
+    return open(
+        resolve_from_site_packages(what),
+        'r'
+    ).read()
